requirements.txt
discord.py==2.4.0
yt-dlp==2024.10.7
ffmpeg==1.4